Don't use redis cache
